requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
Pillow==10.4.0
django-filter==24.3
python-dotenv==1.0.1
channels==4.0.0
daphne==4.1.2
openpyxl==3.1.2